requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Computer Vision
opencv-python-headless==4.10.0.84
mediapipe==0.10.20

# Async HTTP
aiohttp==3.11.11

# Configuration
pydantic==2.10.4
pydantic-settings==2.7.0
pyyaml==6.0.2

# Utilities
numpy>=1.23.5,<2
